requirements.txt
package_name==0.1
packaging==24.2
pillow==11.1.0
pytesseract==0.3.13
